GUI and math
Add the font for the GUI, and improve the code so now it take the
current values from math script.

Add a function in math script that takes the initial values if var
values are null and the currents values if not. The function may be
called from other functions used to change values by player or IA
actions